[#996] Rename crate bump_allocator in iceoryx2-cal to shm_bump_allocator
