+++ iceoryx2/src/service/ipc.rs
@@ -34,8 +34,8 @@
 
 use core::fmt::Debug;
 use iceoryx2_bb_elementary_traits::testing::abandonable::Abandonable;
-use iceoryx2_cal::shm_allocator::bump_allocator::BumpAllocator;
 use iceoryx2_cal::shm_allocator::pool_allocator::PoolAllocator;
+use iceoryx2_cal::shm_allocator::shm_bump_allocator::BumpAllocator;
 use iceoryx2_cal::*;
 
 /// Defines a zero copy inter-process communication setup based on posix mechanisms.
