+++ iceoryx2/src/service/local_threadsafe.rs
@@ -35,8 +35,8 @@
 use core::fmt::Debug;
 
 use iceoryx2_bb_elementary_traits::testing::abandonable::Abandonable;
-use iceoryx2_cal::shm_allocator::bump_allocator::BumpAllocator;
 use iceoryx2_cal::shm_allocator::pool_allocator::PoolAllocator;
+use iceoryx2_cal::shm_allocator::shm_bump_allocator::BumpAllocator;
 use iceoryx2_cal::*;
 
 /// Defines a process local or single address space communication setup.
